pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tightwad"
version = "0.1.0"
description = "Mixed-vendor GPU inference cluster manager with speculative decoding"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "httpx>=0.25",
    "starlette>=0.36",
    "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "httpx[http2]>=0.25"]

[project.scripts]
tightwad = "tightwad.cli:cli"

[tool.setuptools.packages.find]
include = ["tightwad*"]
